+++ mux/src/attrcache.h
@@ -1,6 +1,6 @@
 // attrcache.h
 //
-// $Id: attrcache.h,v 1.3 2003-02-06 14:14:20 sdennis Exp $
+// $Id: attrcache.h,v 1.4 2004-07-07 18:34:35 sdennis Exp $
 //
 
 #ifndef _ATTRCACHE_H
@@ -13,7 +13,7 @@ typedef struct Aname
 } Aname;
 
 extern const char *cache_get(Aname *nam, int *pLen);
-extern bool cache_put(Aname *nam, const char *obj, int len);
+extern bool cache_put(Aname *nam, const char *obj, size_t len);
 extern int  cache_init(const char *game_dir_file, const char *game_pag_file,
     int nCachePages);
 extern void cache_close(void);
